Changed how the overall score is calculated.
We should use the scores calculated by the AI which
range from 1 to 10 as we instructed to do so.

The parameter rating is the BGG rating, which is not
calculated by the AI.

Likewise the fields for cost, duration and players should not be taken into account when calculating a rating.

Only use the fields: fun, learning_curve and difficulty.

Fixes Top 100 includes very high/low scores #35